+++ tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from wtforms.fields.choices import SelectChoice
+from wtforms.fields.choices import Choice
 from wtforms.i18n import DummyTranslations
 
 
@@ -28,8 +28,8 @@ def basic_widget_dummy_field(dummy_field_class):
 def select_dummy_field(dummy_field_class):
     return dummy_field_class(
         [
-            SelectChoice("foo", "lfoo", _selected=True),
-            SelectChoice("bar", "lbar", _selected=False),
+            Choice("foo", "lfoo", selected=True, render_kw={}),
+            Choice("bar", "lbar", selected=False, render_kw={}),
         ]
     )
 
@@ -56,6 +56,7 @@ class DummyField:
         label=None,
         id=None,
         field_type="StringField",
+        groups=None,
     ):
         self.data = data
         self.name = name
@@ -64,6 +65,7 @@ class DummyField:
         self.label = label
         self.id = id if id else ""
         self.type = field_type
+        self._groups = groups
 
     def __call__(self, **other):
         return self.data
@@ -80,11 +82,17 @@ class DummyField:
     def iter_choices(self):
         return iter(self.data)
 
-    def iter_groups(self):
-        return []
+    def _iter_choices_normalized(self):
+        return iter(self.data)
 
     def has_groups(self):
-        return False
+        return self._groups is not None
+
+    def iter_groups(self):
+        return iter(self._groups or [])
+
+    def _iter_groups_normalized(self):
+        return iter(self._groups or [])
 
     def gettext(self, string):
         return self._translations.gettext(string)