+++ tests/fields/test_radio.py
@@ -1,7 +1,7 @@
 from tests.common import DummyPostData
 from wtforms import validators
 from wtforms.fields import RadioField
-from wtforms.fields.choices import Choice
+from wtforms.fields.choices import SelectChoice
 from wtforms.form import Form
 
 
@@ -10,10 +10,14 @@ def make_form(name="F", **fields):
 
 
 class F(Form):
-    a = RadioField(choices=[Choice("a", "hello"), Choice("b", "bye")], default="a")
-    b = RadioField(choices=[Choice(1, "Item 1"), Choice(2, "Item 2")], coerce=int)
+    a = RadioField(
+        choices=[SelectChoice("a", "hello"), SelectChoice("b", "bye")], default="a"
+    )
+    b = RadioField(
+        choices=[SelectChoice(1, "Item 1"), SelectChoice(2, "Item 2")], coerce=int
+    )
     c = RadioField(
-        choices=[Choice("a", "Item 1"), Choice("b", "Item 2")],
+        choices=[SelectChoice("a", "Item 1"), SelectChoice("b", "Item 2")],
         validators=[validators.InputRequired()],
     )
 
@@ -55,7 +59,7 @@ def test_text_coercion():
     # Regression test for text coercion scenarios where the value is a boolean.
     F = make_form(
         a=RadioField(
-            choices=[Choice(True, "yes"), Choice(False, "no")],
+            choices=[SelectChoice(True, "yes"), SelectChoice(False, "no")],
             coerce=lambda x: False if x == "False" else bool(x),
         )
     )
@@ -72,7 +76,7 @@ def test_text_coercion():
 
 def test_callable_choices():
     def choices():
-        return [Choice("a", "hello"), Choice("b", "bye")]
+        return [SelectChoice("a", "hello"), SelectChoice("b", "bye")]
 
     class F(Form):
         a = RadioField(choices=choices, default="a")
@@ -113,7 +117,7 @@ def test_required_validator():
 def test_render_kw_preserved():
     F = make_form(
         a=RadioField(
-            choices=[Choice(True, "yes"), Choice(False, "no")],
+            choices=[SelectChoice(True, "yes"), SelectChoice(False, "no")],
             render_kw=dict(disabled=True),
         )
     )